refactor(maniphest): Convert module to package structure
Split maniphest.py (3,317 lines) into focused submodules:
- core.py: Main Maniphest class orchestrator
- utils.py: Utility functions (timestamps, variable rendering)
- validators.py: Input validation
- resolvers.py: PHID resolution
- fetchers.py: API data fetching
- filters.py: Task filtering
- formatters.py: Data formatting

- Remove underscore prefix from utility functions now at module level
- Update test imports to use new package structure
- Maintain backward compatibility via __init__.py re-exports
