+++ tests/test_maniphest.py
@@ -12,46 +12,45 @@ from phabfive.exceptions import (
 )
 
 # phabfive imports
-from phabfive.maniphest import (
-    Maniphest,
-    _build_dependency_graph,
-    _detect_circular_dependencies,
-    _extract_variable_dependencies,
-    _render_variables_with_dependency_resolution,
-    _topological_sort,
-    parse_time_with_unit,
+from phabfive.maniphest import Maniphest
+from phabfive.maniphest.utils import (
+    build_dependency_graph,
+    detect_circular_dependencies,
+    extract_variable_dependencies,
+    topological_sort,
+    render_variables_with_dependency_resolution,
 )
-from phabfive.column_transitions import (
-    ColumnPattern,
+from phabfive.maniphest_transitions import (
+    TransitionPattern,
     _parse_single_condition,
-    parse_column_patterns,
+    parse_transition_patterns,
 )
 
 
 class TestExtractVariableDependencies:
     def test_simple_variable_reference(self):
-        result = _extract_variable_dependencies("{{ foo }}")
+        result = extract_variable_dependencies("{{ foo }}")
         assert result == {"foo"}
 
     def test_multiple_variables(self):
-        result = _extract_variable_dependencies("{{ foo }} and {{ bar }}")
+        result = extract_variable_dependencies("{{ foo }} and {{ bar }}")
         assert result == {"foo", "bar"}
 
     def test_variable_with_filter(self):
-        result = _extract_variable_dependencies("{{ foo | upper }}")
+        result = extract_variable_dependencies("{{ foo | upper }}")
         assert result == {"foo"}
 
     def test_variable_with_expression(self):
-        result = _extract_variable_dependencies("{{ foo + bar }}")
+        result = extract_variable_dependencies("{{ foo + bar }}")
         assert result == {"foo", "bar"}
 
     def test_no_variables(self):
-        result = _extract_variable_dependencies("plain text")
+        result = extract_variable_dependencies("plain text")
         assert result == set()
 
     def test_invalid_template(self):
         # Invalid syntax should return empty set with warning
-        result = _extract_variable_dependencies("{{ unclosed")
+        result = extract_variable_dependencies("{{ unclosed")
         assert result == set()
 
 
@@ -62,7 +61,7 @@ class TestBuildDependencyGraph:
             "b": "{{ a }}",
             "c": "{{ b }}",
         }
-        graph = _build_dependency_graph(variables)
+        graph = build_dependency_graph(variables)
         assert graph == {
             "a": set(),
             "b": {"a"},
@@ -75,7 +74,7 @@ class TestBuildDependencyGraph:
             "b": "value_b",
             "c": "{{ a }} and {{ b }}",
         }
-        graph = _build_dependency_graph(variables)
+        graph = build_dependency_graph(variables)
         assert graph == {
             "a": set(),
             "b": set(),
@@ -88,7 +87,7 @@ class TestBuildDependencyGraph:
             "b": ["list"],
             "c": "{{ a }}",
         }
-        graph = _build_dependency_graph(variables)
+        graph = build_dependency_graph(variables)
         assert graph == {
             "a": set(),
             "b": set(),
@@ -99,14 +98,14 @@ class TestBuildDependencyGraph:
         variables = {
             "a": "{{ nonexistent }}",
         }
-        graph = _build_dependency_graph(variables)
+        graph = build_dependency_graph(variables)
         assert graph == {
             "a": set(),  # nonexistent is filtered out
         }
 
     def test_empty_variables(self):
         variables = {}
-        graph = _build_dependency_graph(variables)
+        graph = build_dependency_graph(variables)
         assert graph == {}
 
 
@@ -117,7 +116,7 @@ class TestDetectCircularDependencies:
             "b": {"a"},
             "c": {"b"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is False
         assert cycle_path == []
 
@@ -126,7 +125,7 @@ class TestDetectCircularDependencies:
             "a": {"b"},
             "b": {"a"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is True
         assert len(cycle_path) > 0
         # Cycle should contain both nodes
@@ -137,7 +136,7 @@ class TestDetectCircularDependencies:
         graph = {
             "a": {"a"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is True
         assert "a" in cycle_path
 
@@ -147,7 +146,7 @@ class TestDetectCircularDependencies:
             "b": {"c"},
             "c": {"a"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is True
         assert len(cycle_path) > 0
         # All three nodes should be in the cycle
@@ -162,7 +161,7 @@ class TestDetectCircularDependencies:
             "c": {"d"},
             "d": {"a"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is True
         assert len(cycle_path) >= 4
 
@@ -173,7 +172,7 @@ class TestDetectCircularDependencies:
             "c": set(),
             "d": {"c"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is False
         assert cycle_path == []
 
@@ -184,7 +183,7 @@ class TestDetectCircularDependencies:
             "c": {"d"},
             "d": {"c"},
         }
-        has_cycle, cycle_path = _detect_circular_dependencies(graph)
+        has_cycle, cycle_path = detect_circular_dependencies(graph)
         assert has_cycle is True
         # Cycle should contain c and d
         assert "c" in cycle_path
@@ -198,7 +197,7 @@ class TestTopologicalSort:
             "b": set(),
             "c": set(),
         }
-        result = _topological_sort(graph)
+        result = topological_sort(graph)
         # All variables should be present
         assert set(result) == {"a", "b", "c"}
 
@@ -208,7 +207,7 @@ class TestTopologicalSort:
             "b": {"a"},
             "c": {"b"},
         }
-        result = _topological_sort(graph)
+        result = topological_sort(graph)
         # a must come before b, b must come before c
         assert result.index("a") < result.index("b")
         assert result.index("b") < result.index("c")
@@ -220,7 +219,7 @@ class TestTopologicalSort:
             "c": {"a"},
             "d": {"b", "c"},
         }
-        result = _topological_sort(graph)
+        result = topological_sort(graph)
         # a must come before b, c, and d
         assert result.index("a") < result.index("b")
         assert result.index("a") < result.index("c")
@@ -236,7 +235,7 @@ class TestTopologicalSort:
             "c": set(),
             "d": {"c"},
         }
-        result = _topological_sort(graph)
+        result = topological_sort(graph)
         # Within each chain, order must be preserved
         assert result.index("a") < result.index("b")
         assert result.index("c") < result.index("d")
@@ -249,7 +248,7 @@ class TestTopologicalSort:
             "leaf1": {"branch1"},
             "leaf2": {"branch1", "branch2"},
         }
-        result = _topological_sort(graph)
+        result = topological_sort(graph)
         # root must come first
         assert result.index("root") < result.index("branch1")
         assert result.index("root") < result.index("branch2")
@@ -265,7 +264,7 @@ class TestRenderVariablesWithDependencyResolution:
             "name": "John",
             "greeting": "Hello {{ name }}",
         }
-        result = _render_variables_with_dependency_resolution(variables)
+        result = render_variables_with_dependency_resolution(variables)
         assert result["name"] == "John"
         assert result["greeting"] == "Hello John"
 
@@ -276,7 +275,7 @@ class TestRenderVariablesWithDependencyResolution:
             "level2": "{{ level1 }}!",
             "level3": "Message: {{ level2 }}",
         }
-        result = _render_variables_with_dependency_resolution(variables)
+        result = render_variables_with_dependency_resolution(variables)
         assert result["base"] == "world"
         assert result["level1"] == "Hello world"
         assert result["level2"] == "Hello world!"
@@ -288,7 +287,7 @@ class TestRenderVariablesWithDependencyResolution:
             "z_greeting": "Hello {{ a_name }}",
             "a_name": "Alice",
         }
-        result = _render_variables_with_dependency_resolution(variables)
+        result = render_variables_with_dependency_resolution(variables)
         assert result["a_name"] == "Alice"
         assert result["z_greeting"] == "Hello Alice"
 
@@ -298,7 +297,7 @@ class TestRenderVariablesWithDependencyResolution:
             "list": [1, 2, 3],
             "text": "plain",
         }
-        result = _render_variables_with_dependency_resolution(variables)
+        result = render_variables_with_dependency_resolution(variables)
         assert result["num"] == 42
         assert result["list"] == [1, 2, 3]
         assert result["text"] == "plain"
@@ -309,7 +308,7 @@ class TestRenderVariablesWithDependencyResolution:
             "b": "{{ a }}",
         }
         with pytest.raises(PhabfiveDataException) as exc_info:
-            _render_variables_with_dependency_resolution(variables)
+            render_variables_with_dependency_resolution(variables)
         assert "Circular reference" in str(exc_info.value)
 
     def test_self_reference_raises_exception(self):
@@ -317,7 +316,7 @@ class TestRenderVariablesWithDependencyResolution:
             "a": "{{ a }}",
         }
         with pytest.raises(PhabfiveDataException) as exc_info:
-            _render_variables_with_dependency_resolution(variables)
+            render_variables_with_dependency_resolution(variables)
         assert "Circular reference" in str(exc_info.value)
 
     def test_indirect_circular_reference_raises_exception(self):
@@ -327,7 +326,7 @@ class TestRenderVariablesWithDependencyResolution:
             "c": "{{ a }}",
         }
         with pytest.raises(PhabfiveDataException) as exc_info:
-            _render_variables_with_dependency_resolution(variables)
+            render_variables_with_dependency_resolution(variables)
         assert "Circular reference" in str(exc_info.value)
         # Error message should show the cycle path
         error_msg = str(exc_info.value)
@@ -344,7 +343,7 @@ class TestRenderVariablesWithDependencyResolution:
             "port": 443,
             "full_address": "{{ base_url }}:{{ port }}",
         }
-        result = _render_variables_with_dependency_resolution(variables)
+        result = render_variables_with_dependency_resolution(variables)
         assert result["env"] == "production"
         assert result["domain"] == "example.com"
         assert result["protocol"] == "https"
@@ -356,7 +355,7 @@ class TestRenderVariablesWithDependencyResolution:
 
 
 class TestFetchProjectNamesForBoards:
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_fetch_project_names(self, mock_init):
         mock_init.return_value = None
         maniphest = Maniphest()
@@ -385,7 +384,7 @@ class TestFetchProjectNamesForBoards:
         result = maniphest._fetch_project_names_for_boards(tasks_data)
         assert result == {"PHID-PROJ-123": "Project A", "PHID-PROJ-456": "Project B"}
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_fetch_project_names_with_list_boards(self, mock_init):
         """Test handling when boards is a list (empty or non-empty)."""
         mock_init.return_value = None
@@ -410,7 +409,7 @@ class TestFetchProjectNamesForBoards:
 class TestBuildTaskBoards:
     """Test that _build_task_boards creates correct data structures."""
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_build_with_empty_project_names(self, mock_init):
         mock_init.return_value = None
         maniphest = Maniphest()
@@ -423,7 +422,7 @@ class TestBuildTaskBoards:
         assert "Unknown" in result
         assert result["Unknown"]["Column"] == "Backlog"  # First column only
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_build_with_project_names(self, mock_init):
         mock_init.return_value = None
         maniphest = Maniphest()
@@ -437,7 +436,7 @@ class TestBuildTaskBoards:
         assert "Project A" in result
         assert result["Project A"]["Column"] == "Backlog"
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_build_with_list_boards(self, mock_init):
         """Test that list boards format returns empty dict."""
         mock_init.return_value = None
@@ -449,7 +448,7 @@ class TestBuildTaskBoards:
         # Should return empty dict
         assert result == {}
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_build_with_list_boards_single_project(self, mock_init):
         """Test that list boards format returns empty dict."""
         mock_init.return_value = None
@@ -502,12 +501,12 @@ class TestParseSingleCondition:
     def test_invalid_type(self):
         with pytest.raises(PhabfiveException) as exc:
             _parse_single_condition("invalid:Column")
-        assert "Invalid column condition type" in str(exc.value)
+        assert "Invalid transition condition type" in str(exc.value)
 
     def test_missing_colon(self):
         with pytest.raises(PhabfiveException) as exc:
             _parse_single_condition("notavalidpattern")
-        assert "Invalid column condition syntax" in str(exc.value)
+        assert "Invalid transition condition syntax" in str(exc.value)
 
     def test_empty_column_name(self):
         with pytest.raises(PhabfiveException) as exc:
@@ -554,15 +553,15 @@ class TestParseSingleCondition:
         assert result == {"type": "forward", "negated": True}
 
 
-class TestParseColumnPatterns:
+class TestParseTransitionPatterns:
     def test_single_simple_pattern(self):
-        patterns = parse_column_patterns("backward")
+        patterns = parse_transition_patterns("backward")
         assert len(patterns) == 1
         assert len(patterns[0].conditions) == 1
         assert patterns[0].conditions[0]["type"] == "backward"
 
     def test_single_from_pattern(self):
-        patterns = parse_column_patterns("from:In Progress:forward")
+        patterns = parse_transition_patterns("from:In Progress:forward")
         assert len(patterns) == 1
         assert patterns[0].conditions[0] == {
             "type": "from",
@@ -571,20 +570,20 @@ class TestParseColumnPatterns:
         }
 
     def test_or_patterns_with_comma(self):
-        patterns = parse_column_patterns("backward,to:Done")
+        patterns = parse_transition_patterns("backward,to:Done")
         assert len(patterns) == 2
         assert patterns[0].conditions[0]["type"] == "backward"
         assert patterns[1].conditions[0] == {"type": "to", "column": "Done"}
 
     def test_and_patterns_with_plus(self):
-        patterns = parse_column_patterns("from:In Progress+in:Done")
+        patterns = parse_transition_patterns("from:In Progress+in:Done")
         assert len(patterns) == 1
         assert len(patterns[0].conditions) == 2
         assert patterns[0].conditions[0] == {"type": "from", "column": "In Progress"}
         assert patterns[0].conditions[1] == {"type": "in", "column": "Done"}
 
     def test_complex_or_and_combination(self):
-        patterns = parse_column_patterns("from:A:forward+in:B,to:C")
+        patterns = parse_transition_patterns("from:A:forward+in:B,to:C")
         assert len(patterns) == 2
         # First pattern: from:A:forward AND in:B
         assert len(patterns[0].conditions) == 2
@@ -593,17 +592,17 @@ class TestParseColumnPatterns:
 
     def test_empty_pattern_error(self):
         with pytest.raises(PhabfiveException) as exc:
-            parse_column_patterns("")
+            parse_transition_patterns("")
         assert "Empty transition pattern" in str(exc.value)
 
     def test_whitespace_handling(self):
-        patterns = parse_column_patterns(" from:A , to:B ")
+        patterns = parse_transition_patterns(" from:A , to:B ")
         assert len(patterns) == 2
 
 
-class TestColumnPatternMatching:
+class TestTransitionPatternMatching:
     def test_matches_backward(self):
-        pattern = ColumnPattern([{"type": "backward"}])
+        pattern = TransitionPattern([{"type": "backward"}])
 
         # Transaction with backward movement (sequence 2 -> 1)
         transactions = [
@@ -620,7 +619,7 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, None, column_info) is True
 
     def test_matches_forward(self):
-        pattern = ColumnPattern([{"type": "forward"}])
+        pattern = TransitionPattern([{"type": "forward"}])
 
         # Transaction with forward movement (sequence 1 -> 2)
         transactions = [
@@ -637,7 +636,7 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, None, column_info) is True
 
     def test_matches_from_forward_direction(self):
-        pattern = ColumnPattern(
+        pattern = TransitionPattern(
             [{"type": "from", "column": "Backlog", "direction": "forward"}]
         )
 
@@ -655,13 +654,13 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, None, column_info) is True
 
     def test_matches_in(self):
-        pattern = ColumnPattern([{"type": "in", "column": "Blocked"}])
+        pattern = TransitionPattern([{"type": "in", "column": "Blocked"}])
 
         assert pattern.matches([], "Blocked", {}) is True
         assert pattern.matches([], "Done", {}) is False
 
     def test_matches_to(self):
-        pattern = ColumnPattern([{"type": "to", "column": "Done"}])
+        pattern = TransitionPattern([{"type": "to", "column": "Done"}])
 
         transactions = [
             {
@@ -677,7 +676,7 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, None, column_info) is True
 
     def test_matches_been(self):
-        pattern = ColumnPattern([{"type": "been", "column": "Blocked"}])
+        pattern = TransitionPattern([{"type": "been", "column": "Blocked"}])
 
         transactions = [
             {
@@ -693,7 +692,7 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, None, column_info) is True
 
     def test_matches_never(self):
-        pattern = ColumnPattern([{"type": "never", "column": "Blocked"}])
+        pattern = TransitionPattern([{"type": "never", "column": "Blocked"}])
 
         transactions = [
             {
@@ -710,7 +709,7 @@ class TestColumnPatternMatching:
 
     def test_matches_and_conditions(self):
         # Pattern with AND: from:A AND in:B
-        pattern = ColumnPattern(
+        pattern = TransitionPattern(
             [
                 {"type": "from", "column": "In Progress"},
                 {"type": "in", "column": "Done"},
@@ -733,7 +732,7 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, "Blocked", column_info) is False
 
     def test_no_match(self):
-        pattern = ColumnPattern([{"type": "to", "column": "Done"}])
+        pattern = TransitionPattern([{"type": "to", "column": "Done"}])
 
         transactions = [
             {
@@ -749,10 +748,10 @@ class TestColumnPatternMatching:
         assert pattern.matches(transactions, None, column_info) is False
 
 
-class TestColumnFilteringIntegration:
-    """Integration tests for the full column filtering workflow."""
+class TestTransitionFilteringIntegration:
+    """Integration tests for the full transition filtering workflow."""
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_full_filtering_workflow(self, mock_init):
         """Test the complete flow: search -> fetch transitions -> filter -> display."""
         mock_init.return_value = None
@@ -889,7 +888,7 @@ class TestColumnFilteringIntegration:
         }
 
         # Parse transition pattern: find tasks with backward movement
-        patterns = parse_column_patterns("backward")
+        patterns = parse_transition_patterns("backward")
 
         # Verify pattern parsing
         assert len(patterns) == 1
@@ -948,7 +947,7 @@ class TestColumnFilteringIntegration:
 class TestYAMLOutput:
     """Test that YAML output is properly formatted and parsable."""
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_yaml_output_is_parsable(self, mock_init, capsys):
         """Test that task_search generates valid YAML output."""
         from io import StringIO
@@ -1024,8 +1023,13 @@ class TestYAMLOutput:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        # Call task_search
-        maniphest.task_search(tag="Test Project")
+        # Call task_search to get data
+        result = maniphest.task_search(tag="Test Project")
+
+        # Use display_tasks to render output
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         # Capture output
         captured = capsys.readouterr()
@@ -1065,84 +1069,75 @@ class TestYAMLOutput:
 class TestYAMLQuoting:
     """Test that special characters are properly quoted in YAML output."""
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_colon(self, mock_init):
+    def test_needs_yaml_quoting_colon(self):
         """Test that colons trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("foo:bar") is True
-        assert maniphest._needs_yaml_quoting("http://example.com") is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_braces(self, mock_init):
+        assert _needs_yaml_quoting("foo:bar") is True
+        assert _needs_yaml_quoting("http://example.com") is True
+
+    def test_needs_yaml_quoting_braces(self):
         """Test that curly braces trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("{foo}") is True
-        assert maniphest._needs_yaml_quoting("${variable}") is True
-        assert maniphest._needs_yaml_quoting("foo}bar") is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_brackets(self, mock_init):
+        assert _needs_yaml_quoting("{foo}") is True
+        assert _needs_yaml_quoting("${variable}") is True
+        assert _needs_yaml_quoting("foo}bar") is True
+
+    def test_needs_yaml_quoting_brackets(self):
         """Test that square brackets trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("[BUG]") is True
-        assert maniphest._needs_yaml_quoting("[FEATURE] Add something") is True
-        assert maniphest._needs_yaml_quoting("foo]bar") is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_backticks(self, mock_init):
+        assert _needs_yaml_quoting("[BUG]") is True
+        assert _needs_yaml_quoting("[FEATURE] Add something") is True
+        assert _needs_yaml_quoting("foo]bar") is True
+
+    def test_needs_yaml_quoting_backticks(self):
         """Test that backticks trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("`code`") is True
-        assert maniphest._needs_yaml_quoting("Run `make build`") is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_single_quotes(self, mock_init):
+        assert _needs_yaml_quoting("`code`") is True
+        assert _needs_yaml_quoting("Run `make build`") is True
+
+    def test_needs_yaml_quoting_single_quotes(self):
         """Test that single quotes trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("'LOREM'") is True
-        assert maniphest._needs_yaml_quoting("It's working") is True
-        assert maniphest._needs_yaml_quoting("Don't do that") is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_double_quotes(self, mock_init):
+        assert _needs_yaml_quoting("'LOREM'") is True
+        assert _needs_yaml_quoting("It's working") is True
+        assert _needs_yaml_quoting("Don't do that") is True
+
+    def test_needs_yaml_quoting_double_quotes(self):
         """Test that double quotes trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting('"LOREM"') is True
-        assert maniphest._needs_yaml_quoting('Say "hello"') is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_empty_string(self, mock_init):
+        assert _needs_yaml_quoting('"LOREM"') is True
+        assert _needs_yaml_quoting('Say "hello"') is True
+
+    def test_needs_yaml_quoting_empty_string(self):
         """Test that empty strings trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("") is True
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_safe_strings(self, mock_init):
+        assert _needs_yaml_quoting("") is True
+
+    def test_needs_yaml_quoting_safe_strings(self):
         """Test that safe strings don't trigger quoting."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting("Normal task name") is False
-        assert maniphest._needs_yaml_quoting("Task with numbers 123") is False
-        assert maniphest._needs_yaml_quoting("Task-with-dashes") is False
-        assert maniphest._needs_yaml_quoting("Task_with_underscores") is False
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
-    def test_needs_yaml_quoting_non_string(self, mock_init):
+        assert _needs_yaml_quoting("Normal task name") is False
+        assert _needs_yaml_quoting("Task with numbers 123") is False
+        assert _needs_yaml_quoting("Task-with-dashes") is False
+        assert _needs_yaml_quoting("Task_with_underscores") is False
+
+    def test_needs_yaml_quoting_non_string(self):
         """Test that non-strings return False."""
-        mock_init.return_value = None
-        maniphest = Maniphest()
-        assert maniphest._needs_yaml_quoting(123) is False
-        assert maniphest._needs_yaml_quoting(None) is False
-        assert maniphest._needs_yaml_quoting(["list"]) is False
+        from phabfive.cli import _needs_yaml_quoting
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+        assert _needs_yaml_quoting(123) is False
+        assert _needs_yaml_quoting(None) is False
+        assert _needs_yaml_quoting(["list"]) is False
+
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_yaml_output_with_brackets(self, mock_init, capsys):
         """Test that task names with square brackets produce valid YAML."""
         from io import StringIO
@@ -1182,7 +1177,11 @@ class TestYAMLQuoting:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1194,7 +1193,7 @@ class TestYAMLQuoting:
         assert len(parsed_data) == 1
         assert parsed_data[0]["Task"]["Name"] == "[BUG] Something is broken"
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_yaml_output_with_backticks(self, mock_init, capsys):
         """Test that task names with backticks produce valid YAML."""
         from io import StringIO
@@ -1234,7 +1233,11 @@ class TestYAMLQuoting:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1247,7 +1250,7 @@ class TestYAMLQuoting:
         assert parsed_data[0]["Task"]["Name"] == "Fix `make build` command"
         assert parsed_data[0]["Task"]["Description"] == "Run `make test` first"
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_yaml_output_with_mixed_special_chars(self, mock_init, capsys):
         """Test task with multiple special character types."""
         from io import StringIO
@@ -1287,7 +1290,11 @@ class TestYAMLQuoting:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1302,7 +1309,7 @@ class TestYAMLQuoting:
             == "[BUG]: Fix {template} rendering in `parser.py`"
         )
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_yaml_output_with_single_quotes(self, mock_init, capsys):
         """Test that task names with single quotes produce valid YAML with preserved quotes."""
         from io import StringIO
@@ -1342,7 +1349,11 @@ class TestYAMLQuoting:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1355,7 +1366,7 @@ class TestYAMLQuoting:
         # Single quotes should be preserved in the parsed value
         assert parsed_data[0]["Task"]["Name"] == "'LOREM'"
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_yaml_output_with_double_quotes(self, mock_init, capsys):
         """Test that task names with double quotes produce valid YAML with preserved quotes."""
         from io import StringIO
@@ -1395,7 +1406,11 @@ class TestYAMLQuoting:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1412,7 +1427,7 @@ class TestYAMLQuoting:
 class TestStrictFormat:
     """Test that strict format produces guaranteed valid YAML via ruamel.yaml."""
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_strict_format_output_is_valid_yaml(self, mock_init, capsys):
         """Test that strict format produces valid YAML."""
         from io import StringIO
@@ -1456,7 +1471,11 @@ class TestStrictFormat:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1471,7 +1490,7 @@ class TestStrictFormat:
         # Reset to default
         Phabfive.set_output_options(output_format="rich")
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_strict_format_with_special_chars(self, mock_init, capsys):
         """Test that strict format handles special characters correctly."""
         from io import StringIO
@@ -1515,7 +1534,11 @@ class TestStrictFormat:
         mock_response.get.return_value = {"after": None}
         maniphest.phab.maniphest.search.return_value = mock_response
 
-        maniphest.task_search(tag="Test Project")
+        result = maniphest.task_search(tag="Test Project")
+
+        from phabfive.cli import display_tasks
+
+        display_tasks(result, "rich", maniphest)
 
         captured = capsys.readouterr()
         yaml_output = captured.out
@@ -1538,7 +1561,7 @@ class TestStrictFormat:
 class TestTaskSearchTextQuery:
     """Test suite for free-text search functionality."""
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_with_text_query_only(self, mock_init, capsys):
         """Test searching with only a free-text query."""
         mock_init.return_value = None
@@ -1570,7 +1593,7 @@ class TestTaskSearchTextQuery:
         # Should NOT have projects constraint
         assert "projects" not in constraints
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_with_tag_only(self, mock_init, capsys):
         """Test searching with only --tag option (backward compat)."""
         mock_init.return_value = None
@@ -1609,7 +1632,7 @@ class TestTaskSearchTextQuery:
         # Should NOT have query constraint
         assert "query" not in constraints
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_with_text_and_tag(self, mock_init, capsys):
         """Test searching with both text query and tag filter."""
         mock_init.return_value = None
@@ -1647,7 +1670,7 @@ class TestTaskSearchTextQuery:
         assert "projects" in constraints
         assert constraints["projects"] == ["PHID-PROJ-123"]
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_requires_at_least_one_filter(self, mock_init, capsys):
         """Test that search without any filters raises an exception."""
         mock_init.return_value = None
@@ -1661,7 +1684,7 @@ class TestTaskSearchTextQuery:
         # Verify the error message contains helpful information
         assert "No search criteria specified" in str(exc_info.value)
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_with_text_and_date_filters(self, mock_init, capsys):
         """Test text search with date filters."""
         mock_init.return_value = None
@@ -1696,7 +1719,7 @@ class TestTaskSearchTextQuery:
         assert isinstance(constraints["createdStart"], int)
         assert isinstance(constraints["modifiedStart"], int)
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_tag_supports_wildcards(self, mock_init, capsys):
         """Test that --tag option supports wildcard patterns."""
         mock_init.return_value = None
@@ -1732,7 +1755,7 @@ class TestTaskSearchTextQuery:
         # With 2 matching projects, should make 2 calls
         assert maniphest.phab.maniphest.search.call_count >= 2
 
-    @patch("phabfive.maniphest.Phabfive.__init__")
+    @patch("phabfive.maniphest.core.Phabfive.__init__")
     def test_task_search_tag_supports_and_or_logic(self, mock_init, capsys):
         """Test that --tag option supports AND/OR pattern logic."""
         mock_init.return_value = None
@@ -1767,117 +1790,3 @@ class TestTaskSearchTextQuery:
         assert maniphest.phab.maniphest.search.called
         # Should make multiple calls for OR logic
         assert maniphest.phab.maniphest.search.call_count >= 2
-
-
-class TestParseTimeWithUnit:
-    """Test the parse_time_with_unit function."""
-
-    def test_parse_none(self):
-        """Test that None input returns None."""
-        assert parse_time_with_unit(None) is None
-
-    def test_parse_plain_integer(self):
-        """Test parsing plain integer (backward compatibility)."""
-        assert parse_time_with_unit(7) == 7.0
-        assert parse_time_with_unit(1) == 1.0
-        assert parse_time_with_unit(30) == 30.0
-
-    def test_parse_plain_string_number(self):
-        """Test parsing plain string number (backward compatibility)."""
-        assert parse_time_with_unit("7") == 7.0
-        assert parse_time_with_unit("1") == 1.0
-        assert parse_time_with_unit("30") == 30.0
-
-    def test_parse_hours(self):
-        """Test parsing hours."""
-        # 1 hour = 1/24 days
-        assert parse_time_with_unit("1h") == pytest.approx(1 / 24)
-        # 24 hours = 1 day
-        assert parse_time_with_unit("24h") == pytest.approx(1.0)
-        # 12 hours = 0.5 days
-        assert parse_time_with_unit("12h") == pytest.approx(0.5)
-
-    def test_parse_days_with_unit(self):
-        """Test parsing days with explicit 'd' unit."""
-        assert parse_time_with_unit("1d") == 1.0
-        assert parse_time_with_unit("7d") == 7.0
-        assert parse_time_with_unit("30d") == 30.0
-
-    def test_parse_weeks(self):
-        """Test parsing weeks."""
-        assert parse_time_with_unit("1w") == 7.0
-        assert parse_time_with_unit("2w") == 14.0
-        assert parse_time_with_unit("4w") == 28.0
-
-    def test_parse_months(self):
-        """Test parsing months (30 days)."""
-        assert parse_time_with_unit("1m") == 30.0
-        assert parse_time_with_unit("2m") == 60.0
-        assert parse_time_with_unit("6m") == 180.0
-
-    def test_parse_years(self):
-        """Test parsing years (365 days)."""
-        assert parse_time_with_unit("1y") == 365.0
-        assert parse_time_with_unit("2y") == 730.0
-
-    def test_parse_float_values(self):
-        """Test parsing float values."""
-        assert parse_time_with_unit("1.5w") == 10.5
-        assert parse_time_with_unit("0.5m") == 15.0
-        assert parse_time_with_unit("2.5d") == 2.5
-
-    def test_parse_with_whitespace(self):
-        """Test parsing with whitespace between number and unit."""
-        assert parse_time_with_unit("1 w") == 7.0
-        assert parse_time_with_unit("2  m") == 60.0
-        assert parse_time_with_unit("  7d  ") == 7.0
-
-    def test_parse_case_insensitive(self):
-        """Test that units are case insensitive."""
-        assert parse_time_with_unit("1W") == 7.0
-        assert parse_time_with_unit("1M") == 30.0
-        assert parse_time_with_unit("1Y") == 365.0
-        assert parse_time_with_unit("1D") == 1.0
-        assert parse_time_with_unit("1H") == pytest.approx(1 / 24)
-
-    def test_parse_zero(self):
-        """Test parsing zero values."""
-        assert parse_time_with_unit(0) == 0.0
-        assert parse_time_with_unit("0") == 0.0
-        assert parse_time_with_unit("0d") == 0.0
-        assert parse_time_with_unit("0w") == 0.0
-
-    def test_invalid_empty_string(self):
-        """Test that empty string raises ValueError."""
-        with pytest.raises(ValueError, match="Time value cannot be empty"):
-            parse_time_with_unit("")
-        with pytest.raises(ValueError, match="Time value cannot be empty"):
-            parse_time_with_unit("   ")
-
-    def test_invalid_unit(self):
-        """Test that invalid unit raises ValueError."""
-        with pytest.raises(ValueError, match="Invalid time unit: 'x'"):
-            parse_time_with_unit("5x")
-        with pytest.raises(ValueError, match="Invalid time unit: 'days'"):
-            parse_time_with_unit("5days")
-
-    def test_invalid_format(self):
-        """Test that invalid format raises ValueError."""
-        with pytest.raises(ValueError, match="Invalid time format"):
-            parse_time_with_unit("abc")
-        with pytest.raises(ValueError, match="Invalid time format"):
-            parse_time_with_unit("w1")
-        with pytest.raises(ValueError, match="Invalid time format"):
-            parse_time_with_unit("1.2.3d")
-
-    def test_negative_value(self):
-        """Test that negative values raise ValueError."""
-        with pytest.raises(ValueError, match="Time value cannot be negative"):
-            parse_time_with_unit("-1d")
-        with pytest.raises(ValueError, match="Time value cannot be negative"):
-            parse_time_with_unit("-5w")
-
-    def test_examples_from_issue(self):
-        """Test examples from GitHub issue #105."""
-        # The issue specifically mentions --updated-after=1w
-        assert parse_time_with_unit("1w") == 7.0
